feat(Structure): ensure vrtk scripts are decoupled from sdk rig setup
Previously, the scripts within VRTK would need to be applied to the SDK
rig setup, so for SteamVR the scripts would be required to be placed on
the CameraRig GameObject for affecting the Play Area, or placed on the
specific CameraRig controller for providing functionality to either of
the controllers.

This has now been changed that the CameraRig prefab is now completely
decoupled from any of the VRTK scripts and all of the scripts can (and
should) be placed on separate game objects. The SDK Manager script is
also required in all scenes as it also provides linking of the SDK rig
objects to the VRTK scripts along with linking the SDK controllers to
the game objects that the VRTK controller scripts reside on.

This new set up is very useful if the rig setup for the SDK is
different as it means the script set up doesn't need to change, only
the linked objects in the SDK Manager need to change.

The SDK Manager has been updated as well so it is now a singleton that
provides stateless access to the SDK bridge to determine which SDK
should be used.

All of the example scenes have been updated now to have a clean SteamVR
CameraRig and a separate collection of game objects to hold the VRTK
scripts.